requirements.txt
Django==1.11
pytz==2017.2
psycopg2==2.7.1
django-widget-tweaks==1.4.1
docutils==0.13.1
Pillow==4.1.1
appdirs==1.4.3
django-feedparser==0.2.1
django-jet==1.0.6
feedparser==5.2.1
google-api-python-client==1.6.2
httplib2==0.10.3
oauth2client==4.0.0
packaging==16.8
pyasn1==0.2.3
pyasn1-modules==0.0.8
pyparsing==2.2.0
pytz==2017.2
requests==2.13.0
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
uWSGI==2.0.15
djangorestframework==3.6.2
django-debug-toolbar==1.8
coverage
faker
reportlab==3.4.0
raven==6.1.0
newrelic==2.88.0.72